Rename long folder to generic/instrinsic name
Bit easier to understand
